pdfwrite - better identify rectangular clips, and emit as rectangles
Bug #687931 "GS produces large PDF files"

The file uses *many* rectangular clips to draw through, previously pdfwrite
mostly emitted these as a sequence of lines, because only a small class of
clip paths could be identified as rectangular.

This commit levers the existing gx_path_is_rectangle to identify many more
rectangular clip paths, and emit them using the 're' operator instead, which
is more compact on output.

The specimen file now produces 30% smaller PDF files.

This does introduce some differences, in the main these are very small single
pixel or one pixel wide lines. In the case of the Quality Logic test suite, the
'frame' for the tests is drawn through a clip, and one edge is now slightly
wider, since it better matches the width of the other edges this is a
progression. It affects a large number of QL tests.