pdfwrite - implement idioms for PScript 5 artificial bld text
Bug #688007 "cjk bold fonts printed many times"

The artifical emboldening of fonts in the pscript 5 DLL works by drawing
the same text 4 times in slightly different positions. THis causes
pdfwrite to emit 4 sets of text.

Here we implement idiomns for the range of 'show' operations valid in
Pscript5, and if the device supports text rendering modes we draw the text
only once in text rednering mode 2 (fill and then stroke). The Pscript5
code uses 'sBdx' to determine the offset of the text, but this is actually
a constant derived by dividing the target print resolution by 300. We use
a fixed figure derived heuristically instead.

During the work it became apparent that the existing code in pdfwrite to
handle stroked text didn't work properly as it could emit stroke
parameters during the course of a text object, which Acrobat ignores. SO
the code has been altered to exit the text object whenever text is
stroked. This is inefficient as we might not need to emit any stroke
parameters, but it solves the problem, and text rendering modes are not
widely used.

This shows a small progression in pdf-t3-simple.pdf and some small
differences in other files.