pdfwrite - Fix image writing code when converting colour spaces (old CMS code)
Bug #694401 "pdfwrite produces incorrect pdf file"

The command line for this bug specifies -dPDFSETTINGS=/ebook. This sets a
number of parameters, specifically it sets ColorConversionStrategy to sRGB.
The input file contains an image in DeviceGray with 1 bit per component,
the specified colour conversion strategy means we convert this to RGB.

The original image was drawn with a resolution of 200 dpi, which does not
trigger monochrome downsampling in this setup (300 dpi). However when
converted into a colour image it does trigger downsampling as the threshold
for colour images is 72 dpi in this configuration.

The combination of colour conversion and image downsampling seriously
confused the horrific code in pdfwrite, and led to use writing out the wrong
amount of data for the image as well as corrupting the image data as it
passed through the filters.

This patch sorts this out by setting various parameters, and copying/restoring
them as required to initialise the various filters correctly.

For some reason this shows a difference with test file Bug691425.pdf but the
visual appeaance is the same.