Make W/W* "context" survive text operations.
If we have a text operation between a W/W* operator and the path operation that
ends the W/W* context, we end up entering an infinite loop because the W/W*
context prevents us from writing the changes to the graphics state required
to render the text.

To handle this, we now end the W/W* context, and store the dictionary
temporarily while we complete the text operation, then restore W/W* context
after.

It is not clear how this is supposed to interact with the clipping text
rendering modes, but on the basis of solving an infinite loop, I'm committing
this, and will address issues if they arise.

No cluster differences.