Bugs 693828: pull in new URW+ GPL fonts.
Also 693827, 688557, 691909, 690896 and 687297.

Many differences: most files that use the standard font set change. Changes of
note are that in Nimbus L Sans Italic several of the "symbol" glyphs (like
/plus, /plusminus etc) used to be italicised and are now not. This actually
matches the Adobe equivalent font.

Also there are some cases of the /M glyph being narrower than previously, again
this is correct wrt to the Adobe equivalent.