PDF interpreter - set colour nad space at the time the relevant operators execute
In the past we have only set the colour space and colour values for stroke
and fill when we actually perform a stroke or fill operation. At the time
the space and colour was set we would simply store the data.

This can cause problems with a sequence like this:

/CS0 cs 0.2 0.1 0 0  scn
q
....
f
Q

q
....
f
Q

Each time we set the colour and space we then restore back to a point
before that space was set. On the next fill we need to set the space again.
If the colour space is an expensive one to set (eg ICCBased) then we waste
a lot of time.

The 'simple' solution is to set the colour space at the time it is set in
the PDF file. However, this also meant setting the colour values at the
time they are set (or we could get space and colour unsynchronised). This
further led to the discovery that the transparency code can set colours in
a different space to the saved space (eg for an SMask). So it was necessary
to have an additional set of routines, one which simply stores the state
and one which stores and sets it to be the current space/colour.

The time taken for the specimen file is reduced from approximately 4.5
minutes to around 3.5 seconds. In addition the file CATX1101.pdf shows a
progression with pdfwrite (gray drop shadow was missing). Because pdfwrite
no longer swaps colour spaces so frequently with text, text is more often
emitted as a single operation, which leads to smaller files (15% smaller
in one case) but does show some very minor (1 pixel) positioning changes
in text.