pdfwrite - new colour conversion code, fix Indexed Separation/DeviceN
The previous code for converting the base space of a Separation or DeviceN
space didn't cater for the possibility of a /Indexed space. This code now
does.

In addition we do a better job of identifying when a base space needs
converted, there was similar problem here leading to us converting (eg)
CMYK to CMYK.

There are some oddities with RGB->CMYK conversion that I need to discuss
with Michael.