pdfwrite - new colour code
Concretize_color doesn't work the way I thought it did. Instead of returning
components in the original base space, it returns components for the current
ProcessColorModel, this has resulted in the removal of several steps.

Modified the code which writes converted base spaces for Separation and
DeviceN spaces, so that the space can be used for images, as well as for
linework.

The code is now no longer crashes when the ColorConversionStrategy is
CMYK.

There are still some oddities with CIE/RGB to CMYK conversion, there are
some problems with CMYK output with type 4 images. Still need to test
Gray and RGB strategies, and redo LeaveColorUnchanged and UseDeviceIndependentCOlor
with the latest changes.