Moved CUPS filters gstoraster and gstopxl to cups-filters
As Ghostscript's CUPS filters gstoraster and gstopxl need changes to
allow PPD-less printing on IPP printers and these changes introduce
a new dependency on the libcupsfilters library of the cups-filters
project, we have moved these filters to cups-filters now.

In addition, the pdftoraster filter for Poppler-based conversion
of PDF input to CUPS Raster is also not part of Poppler but part
of cups-filters. So we get a more symmetric distribution.

Also maintenance of the filters will get easier this way, especially
duplicate code can move into libcupsfilters.

cups-filters is a free software package hosted by OpenPrinting and it
contains all CUPS filters needed by CUPS under Linux and additional
utilities for CUPS. The filters are not part of the CUPS project at
Apple any more since CUPS 1.6.x.

See
http://www.linuxfoundation.org/collaborate/workgroups/openprinting/cups-filters