Attempt to handle broken object definitions (garbage on stack)
If we have certain kinds of broken object definitions we can end up with
more than one object on the stack when we encounter an endobj (or a new
object definition before encountering an endobj for the previous object).
Previously we would throw an error and ignore the object (define it as null)
now we attempt to recover by removing objects from the stack until we have
a stack of the form 'n n object' where the 'n' represents an integer.

The bug #694443 now runs to completion although the output is empty. Given
how badly broken this file is I don't propose to do any more with it.