add extra error checking of the Trailer dictionary
Bug #694474 the file is badly damaged, and Acrobat won't open it. The only
problem from our perspective is that the error generated causes the
interpreter to exit (because there is no Trailer dict). Added more checks
before use to ensure we don't generate errors with missing (empty) trailer
dictionaries.

A number of other files no longer cause errors when run with the cluster.