Bug 694502: reverse of dictomark for PDFs
The order with which the key/value pairs in a <<>> (apparently) differs between
Postscript and PDF. Postscript pushes the objects onto the stack, then pops
them in pairs, so the order is reversed compared to the order they were
encountered. PDF handles them in the order they were encountered.

This is, of course, only a problem when there are duplicate keys.

Thus in PS "<< /a 1 /a 2>>" the key "/a" ends up with the value "1". In PDF, "/a" gets
the value "2".

No cluster differences.