PDF interpreter run .reusablestreamdecode in 'stopped' to handle errors
Bug #694437 " Infinitely many warnings: **** File has unbalanced q/Q operators (too many Q's) ****"

The infinite recursion is due to not updating the current file position, and
rerunning the same code, which results in the same error.

To handle this we wrap the exeuction of reusablestreamdecode in a 'stopped'
context, so that we cna capture the error, and move the file pointer as
required. The job still throws warnings, but the resulting PDF is the same
as displayed by Acrobat with the original file.