PDF interpreter - increased checking of Outlines, to avoid errors
Bug #694503 "Regression: output pdf file blank starting with 0c8150ab6b2b8ceaa17cc1b1933d523764ffe114"

After attempting to recover from broken objects, it was possible to
throw an error when processing Outlines, because the recovered object
was of the wrong type. This could cause PDF files with broken /Outlines to
fail with pdfwrite, where previously they would write a file, but drop the
/Outlines. (this only occurs wiht pdfwrite, because we ignore Outlines
when rendering)

This commit restores the previous behaviour by checking the type of the returned
object.