PDF interpreter - improve handling of pdfmarks when we have multiple input files
Bug #688542 "Bookmarks Get Lost In PDF"
Bug #688829 "Merging PDF files using gs: outlines and links not updated"

The pdfmarks generated by the PDF interpreter for Outlines and Links use the
page number relative to the start of the current input file. For single
input files this works fine, but when we have multiple input files, we end
up with outlines and links pointing to the wrong place.

To resolve this we need to know, at the start of any PDF file, how many pages
have already been processed, and then we will add this value to the nominal
page number when creating the pdfmarks.

Hin-Tak provided a solution (#688829) that used globaldict, which Alex rejected
for that reason. This commit executes currentpagedevice at the start of every
PDF input file, and stores in in /CumulativePageCount. We can then use this
to generate pdfmarks which are correct.

Even if PDF and PostScript input files are mingled, the PDF-generated pdfmarks
will still be correct.

We won't make any attempt to correct PostScript input, if you are creating a
PostScript program, you should get it right. If you really must merge, then
use pdfwrite to create the individual PDF files, and then merge those.

It is possible that other types of generated pdfmark will need similar
alteration, we'll tackle those as they turn up.