PDF Interpreter - update to 073f460af5bb37edb1849c5d6235048598100437
Bug #694508 "Regression: error writing PDF file starting with
72b397ecb4f5be3fdbf6da4a90a306d259ec3e93"

The commit 073f460af5bb37edb1849c5d6235048598100437 had an error in the
handling of named actions, which I didn't previously have a test case for.
NB the commit in the bug title is incorrect, I'm not clear on why Marcos
thought this was the commit which introduced the error.

This commit fixes that and allows me to test the named actions, which now
work correctly.