PDF interpreter and pdfwrite - better handling of Outlines and Dests
Bug #690253 "'pdfmark destination page' error while splitting a PDF: confusing output and invalid links."

Several parts to this, we had already removed the 'ERROR' from the message
emitted when a DEST or OUT pdfmark pointed to a page outside the valid range.

This commit makes pdfmark produce an error when an Outline has an invalid
range (DEST laready did). We also wrap pdfmark invocations in 'stopped' so
that we can trap the error. This allows us to elide Links and Outlines to
invalid pages (previously the links remained, but did nothing useful).

Since the bug was raised, we now also use the '%d' syntax to split PDF
files. This would still produce invalid links, because the page number
would not be invalid.

We now check the OutputFile when opening a PDF file, and if it contains
'%d' we set a couple of flags which will skip Outline and Dest link creation
as these would never be correct.