Fix bug 694455. Incorrect CTM when rendering SMasks.
The clippath was incorrect due to the CTM. As specified in 7.5.4 of the PDF spec:

The mask's coordinate system is defined by concatenating the transformation
matrix specified by the Matrix entry in the transparency group's form dictionary
(see Section 4.9.1, "Form Dictionaries") with the current transformation matrix
at the moment the soft mask is established in the graphics state with the gs
operator.

This fixes the customer's file and also results in a progression on
comparefiles/Bug693681.pdf where extra data was previously showing.