PS Interpreter - populate the /OutputDevice resource
Bug #689500 "/OutputDevice category is empty"

This commit uses the devices built into the executable to populate the
/OutputDevice resource. The bug thread suggests using devicedict but we
can't do that in gs_res.ps because it isn't set up until later. So we
duplicate the code used to populate devicedict.

According to the PLRM the only required member of the OutputDevice resource
instance is the PageSize, I've chosen to copy the same values as Adobe Acrobat
Distiller, but these can be changed if this proves to be a limitation.

Two additional changes were required, exposed by the Quality Logic test file
09-59.ps

In gdevdevn.c devn_get_params() it is possible for the function to be called
with pequiv_colors set to NULL, this caused a seg fault. Modified the code
so that it doesn't attempt to dereference a NULL pointer.

The file 09-59.ps uses setpagedevice to set each possible instance of the
OutputDevice resources, this can cause problems with some devices as it
doesn't actually send a file. In particular with the mswinpr2 device on
Windows, but it also caused errors on Linux. Since this is a CET file,
gs_cet.ps was modified to delete the OutputDevice resource category and
create a new empty one.

Expected Differences
430-01.ps enumerates all the entries in /OutputDevice and so has a *lot*
more data to print.