update caption.ps so that it does not mark the page on device deactivation
Bug #694517 "additional page with caption.ps"

The EndPage procedure in caption.ps marks the page in the file with the
defined caption regadless of the reason code for the EndPage. This means
that it marks the page even on device deactivation.

Since commit 3894e9c8a3ab125c82048da3bf81515500bb2da0 pdfwrite will emit
pages that are marked, even if no showpage is executed (this is to match
observed Acrobat Distiller behaviour) and this means that an extra 'blank'
page is emitted on device deactivation.

This commit simply modifies caption.ps so that it only writes the caption
if the reason for calling EndPage is *not* device deactivatyion.