Bug 694451: Solve problems with interpolated landscape images.
The portrait image code accumulates pixels into scanlines, and then
plots the whole scanline at once. The landscape image code follows
this lead by accumulating pixels into 'scancolumns' and then plots
the whole column at once.

Unfortunately this accumulation for columns goes wrong in the case
where we are working at < 8 bpp (as the transposition of rows to
columns fails). This causes images with runs of identical pixels in
to be 'squished' vertically.

We work around this here in the < 8 bpp case by repeating the writes
of each pixel so whole bytes are filled. This is suboptimal, but
doing anything else is tricky with the macros the way they are.