Update antidropout downscaler.
The antidropout downscaler is not consistent in the way it detects
'darker' colors; some of the code predicates 'darker' on the color
polarity, other parts of it do not.

Here we make it consistently follow the color polarity.

Also, we correctly take the color polarity from the color space, rather
than from the device. Thanks to Michael Vrhel for making the changes
to make this possible.