PS Interpreter - yet more meddling with VM allocation and setpagedevice
Bug #694525 "GSView 5.0 does not work with GPL Ghostscript 9.08"

Rather horrifyingly GSView uses undocumented internal operations of
Ghostscript. In this case the changes to VM allocation mode during setpagedevice
were being circumvented.

This commit changes the way we detect a setpagedevice during startup, previously
we check that the dictionary contained one key called /.LockSafetyParams
and if so assumed that this meant we were in startup, so we set the VM
allocation mode to global for the benefit of Display PostScxript.

GSView ended up triggering that test, which caused a VM allocation error.

We now look for the key /.LockSafetyParamsGlobal instead. If this is present
we set the VM allocation mode to global and *replace* the key with
/.LockSafetyParams instead. This allows accurate detection of the startup
condition.

No differences expected, GSView 5 seems to work and the original bug still
is fixed.