Address gp_wgetv not finding *any* environment variables unless compiled with GS_NO_UTF8
The string being passed in as the 'name' was a regular 1 byte C char array.
But _wgetenv expects to be given a wide character string to find and
treats the argument as such. This mean environment variables never matched.

This commit mimics the code for the registry variables, and converts the
name to a wchar before passing to _wgetenv.

CLUSTER_UNTESTED