Allow use of '\\' for filenameforall and friends on Windows
The commit to allow recursive directory enumeration f13bfba removed the
code which translated '\\' into '\', which could lead to incorrect filenames
being returned.

This fix restores that functionality, but we still think everyone would be
better advised to use the PostScript '/' separator instead, which works as
well as '\' does.

CLUSTER_UNTESTED