Make the Windows implementation of filenameforall match the Unix version
The Unix version of filenameforall starts in the current working directory
when no directory is specified.

When the pattern length exceeds the available string we should return a
rangecheck error, not an empty string.

When the filename exceeds the available string we should return a
rangecheck error, not a truncated string.

CLUSTER_UNTESTED