Yet more work on NTFS file enumeration
Bug #686853, SaGS comment #9

We now match the pattern template only at the top level of the recursion,
any directories which match the pattern are enumerated, but their entire
contents are listed.

Multiple path separators are now removed at each level. It should now be
impossible to reach a situation where a returned file has '\/' or similar
contained within it.

Illegal characters ('*' and '?') in the path will now cause no matches to
be returned. These are of course valid after the final separator of the
path and will match filename patterns at that level.

I've checked this with various paths utilising multiple path separators,
mixed use of separators, illegal escaped characters in the path and a modest
variety of patterns. All this with COMPILE_INITS=0 and COMPILE_INITS=1 and
in both cases with GS_NO_UTF8=1 and without specifying GS_NO_UTF8. All seems
to be OK in all cases. None of this is tested by the cluster of course.

CLUSTER_UNTESTED