Fix a compiler warning which exhibited a real problem
In gdevpdfp.c remove a redefinition of 'code'. The 'for' loop uses the
outer definition of 'code', but the inner definition is set by
cos_dict_put_c_key_string() This means that the inner definition is never
used, and the outer one used to control the loop is never changed, defeatimg
the purpose.

The other new compiler warnings in pdfwrite are either not valid (some warnings
are incorrect) or benign and fixing them introduces other warnings.