pdfwrite - make the saved gstate stack dynamically allocated instead of static array
Bug #693591 "Infinite loop writing pdf file"
Bug #694531 " "*** File has unbalanced q/Q" infinite loop on seemingly valid file"

The stack of saved gstates in pdfwrite was a statically allocated array of 11
elements. Cairo produces PDF files with insane (and pointless) levels of
Form XObject nesting which can easily break this limit (as well as causing all
PDF consumers to render the files slowly)

This commit replaces the fixed size array with a dynamically allocated array.
For performance reasons we still allocate 11 elements initially, and if we
exhaust the storage, we reallocate 5 elements larger.

This also shows a progression in the test file 'inefficient_tiling.pdf' this
file is produced by LaTeX but I guess it may be using Cairo behind the
scenes.