More work related to VM mode, setpagdevice and startup conditions
This has a lengthy bug tail attached to it, the summary follows, prior
commits of interest are (ordered by date, earliest first):
073f7be1671e7fa2dbbc75984e593d2279bb05d7
939e32ff3c8841507aed80435269457f3caac6f5
774231a7db63439c72313050a750b219491c172f
39b494d41a814e6b27a5828030e58e0b239044fb

Initially the bug report was #687702 "gs8.15 setpagedevice /invalidaccess",
the problem was caused by a 'fix' for WordPerfect files which assumed a local
VM writeable dictionary returned by currentpagedevice. The fix was to create
a local VM copy of the dictioanry, because its acceptable to store global
objects in local VM, but not vice versa. (073f7be1671e7fa2dbbc75984e593d2279bb05d7)

However, if we start with -dSAFER then, quite early in gs_init.ps, we call
/.setsafe (which called /.locksafe) which ends up executing setpagdevice
in order to set /LockSafetyParams. This causes the page device dictionary to
be allocated in local VM. This should not cause any problems, but.....
setpagedevice stores a pointer to the returned dictionary in the graphics
state, the Display PostScript initialisation code insists that all objects
in the 'savedinitialgstate' be in global VM, and we just allocated the dict
in local VM, so we get an error.

This was fixed in setpagedevice by checking for a 1 element dictioanry
with the specific key '/LockSafetyParams', and creating a global VM dict
instead of a local one. (939e32ff3c8841507aed80435269457f3caac6f5)

We then discovered that GSView sends PostScript containing '.locksafe',
which triggers the check in sepagedevice and sets the VM allocation mode
to global, causing an invalidaccess error. GSView should not be using
undocumented internal routines but it was decided to fix this, even though
a release had just been completed. This was done by creating a new routine
'.locksafeglobal' which sets a key called /LockSafetyParamsGlobal and
having setpagedevice check for that key, if discovered it creates a global
VM dictionary and replaces the key with /LockSafetyParams. Finally /.locksafety
was returned to its original state. (774231a7db63439c72313050a750b219491c172f)

A new release was completed, and it then transpired that, despite having
tested gv briefly, under certain conditions it executes a script (pdf2dsc)
which uses pdf2dsc.ps, which calls .setsafe. This didn't seem terribly
important and so the line was removed from the PS file.

Ray pointed out that the .setsafe routine is (minimally) documented in
use.htm, and it seems is important to the Unix community because of the
'SAFER' switch. So this commit restores /.setsafe (and the line in pdf2dsc.ps)
and creates a new /.setsafeglobal. This is *only* executed in gs_init.ps,
its presence is solely for the benefit of DPS so that the initial gstate
is allocated in global VM, and must not be used anywhere else.

Hopefully we can finally say goodbye to all this mess now.