Fix -dUseCIEColor (part of the whole VM allocation problem)
See commit ae930279498a5961fcf5d70ffe86864883609cbc for a lengthy explanation)
of the background to this problem.

The fix for the original problem involved always creating the device dictionary
for setpagdevice in local VM.

Following on from that we discover the following in zdsp1.c

gstate_check_space(i_ctx_t *i_ctx_p, int_gstate *isp, uint space)
{
    /*
     * ****** WORKAROUND ALERT ******
     * This code doesn't check the space of the non-refs, or copy their
     * contents, so it can create dangling references from global VM to
     * local VM.  Because of this, we simply disallow writing into gstates
     * in global VM (including creating them in the first place) if the
     * save level is greater than 0.
     * ****** WORKAROUND ALERT ******

Now, if we execute the Display PostScript .savelocalstate routine, its vital
that the gstate and all its contents be in global VM in order to satisfy
the test in zdps1.c.

So, any calls to setpagedevice before the call to .savelocal must allocate
the page device dictionary in global VM (as a pointer to the device dictionary
is saved in the gstate)

Thus we need a way to tell setpagedevice to allocate its dictionary in global
VM, even though we usually want it in local VM.

This commit adds a somewhat hacky method for this, we insert a specific key
/..StartupGlobal in the page device dictionary. If setpagedevice finds this
key in the dictionary we allocate the dict in global VM, and remove the key.

We then use this key in a couple of places in gs_init.ps where we execute
setpagedevice *before* we've reached .savelocalstate.

This fixes all the myriad problems with allocation states and gives us a way
to add any more that are required.

I've put a comment into gs_init.ps immediately before /.savelocalstate so
at least in the future it will be easier to figure out where this is required.

I've searched all the initialisation files looking to see if there are any
other likely places (the INITFILES string is parsed before we get to
.savelocalstate) and they all look acceptable. None of the routines actually
execute setpagedevice.

No cluster differences expected as none of this is tested