Fix bug 694539.	Reading off the end of the cbuf when dash pattern len is max.
The max length dash pattern (currently 11) was used frequently in this file,
and in at least one case, this tripped over a memmove that is way too large in
the top_up_cbuf because the pointer was past the end of the buffer, and the
unsigned int count was 0xffffffff. The root cause was that the cmd_largest_size
used to set the warn_limit did not account for both the dot_length and the
offset.

I corrected cmd_largest_size, added some comments, and added an error message
and ioerror return to top_up_cbuf.