pdfwrite - more discriminating tests before TrueType font combining
Bug 694538 "Incorrect embedding of fonts"

The test file does not appear to be the regular output of the MS Windows
PostScript driver, it looks like several files have been stripped and then
added together. A process liable to failure at the best of times.

In this case we end up with a large number of incrementally downloaded,
subset, TrueType fonts. Despite the fact that these are the same font and
in many cases contain the same, or compatible, subsets the fonts are
downloaded several times, which is why I believe this to be a combination
of separately created files.

pdfwrite tries to combine such fonts, and does so quite succcesfully in
this case. However, the individual fonts are not all alike. Some information
in the TrueType table is clearly copied from the original font (eg the
numGlyphs value of 3147) and some is not. In particular the maxPoints
value differs between two otherwise compatible subset fonts (one has additional
glyphs). It seems Acrobat doesn't like this, it sometimes raises an error
and does not display any glyphs whose point count exceeds the maxPoints
value.

This commit adds the maxPoints and maxContours values to the gs_type42_data
structure and tracks them. Fonts are not now considered compatible if
these values are different between them.

This will lead to less font merging, but for sensibly created programs
this will not be a problem, as the driver only downloads the font once. For
programs created by jamming files together it will prevent the situation
where we create a font which Acrobat cannot read.

No differences expected.