PDF interpreter - clamp invalid numbers to 0 in Shading Coords
Bug 694542 "Ghostscript failed to convert pdf file to images"

The specimen file contains invalid numbers in the Coords array of every
Shading dictionary. Acrobat (of course) handles this by silently converting
all such numbers to 0.

This commit checks the Coords of the Shading, and if any entry is not a number
(our token parser converts invalid numbers to names) it replaces it with a
xero.

No differences expected