PostScript interpreter - make setpagedevice and EndPage work more like Adobe
Bug 690667 "Changing PageSize during n-up processing by BeginPage/EndPage generates wrong output"

The test job changes setpagedevice during the course of the job, and the
EndPage takes special action depending on the page count. The first problem
is that we don't reset the count of pages in the device, and Acrobat can
be shown to do so on *any* call to setpagedevice (even if the dictionary
is empty). In zsetdevice() we now reset the page count in the current device
so that he page count is reset, even if the device doesn't change.

Secondly we call .trysetparams before we execute the EndPage routine (if
present). The problem here is that .trysetparams actually modifies the
page device dictionary (and hence the device) and so attributes can change
before the EndPage, in particular the PageSize. The /setpagedevice routine
has been modified to move the execution of EndPage much earlier in the
process. Again testing with Adobe Acrobat shows that the EndPage procedure
is executed if the dictionary argument is empty, and indeed even if the
execution of setpagedevice causes an error.

No differences expected. Note this is a potentially destabilising change.