PDF interpreter - synthesize appearance for multiline text widgets
Bug 689761 "Multiline PDF form field does not honor line breaks"

There are several parts to this change. Firstly there was a bug in the code
to detect multiline text widgets, which meant that they were not detected
and instead of flagging a warning and ignoring the widget, we handled it
as if it were a single line. This also meant that we were mishandling combined
fields as tetx fields.

Secondly there is code in pdf-draw.ps which checks to see if we can synthesise
an appearance and if we can it ignores the one provided. Testing with Acrobat
indicates that this is what Acrobat does. However I think this is very wrong
and in testing our code I see about as many files which are better if we
use the appearance stream as are better if we manufacture one. So I've
chosen (for now at least) to use the appearance stream if we're given one.

Finally we now handle multiline text fields. The string is broken up into
multiple strings at the line breaks. The height of the rectangle is divided
by the number of strings, which allows us to space the lines evenly. We
detect and implement the different types of quadding. Note that Acrobat
does not seem to distribute the text the same way we do, but I don't see
any reason to try and mimic it.

Expected Differences
Bug689450.pdf Bug692968.pdf 1503_text_missing .pdf Bug690526.pdf Bug692591.pdf
these files have appearance streams that do not match up with the field values.
1783_CIDToGIDMap...pdf x_-_complex_fomr.pdf using the appearance stream
improves the output because our synthesis is not that great.
comb_and_password.pdf uses a combined field which we previously treated as
text, and which we now ignore.