PS Interpreter - don't interpolate imagemask data for high level devices
Bug #689460 "imagemask interpolation differs from Adobe"

This is rather a long-running bug. The title isn't really very descriptive
of the problem, as far as I can tell the problem only affects pdfwrite.

In the current implementation, interpolation of imagemasks is handled in
*PostScript*, by inserting the ImScaleDecode filter in front of the image
DataSource. This works well for rendering, but for high level devicees
(particularly pdfwrite) this means that we do not embed the original mask
data, we embed the interpolated data. This makes the PDF file much larger
and also means that when opened with Acrobat Reader, the output does not
match, becuase Acrobat does interpolation by applying what looks like a
Mitchell filter, and promoting the imagemask to 8 bpc.

I did look into applying our current scheme in C, as suggested in comment #18,
but that was inordintely difficult. We can't use the existing 1bpp image
interpolation code, as that promotes to 8bpc and uses a Mitchell filter,
so we lose the sharp edges, which is important for a mask. It probably would
be possible to replicate the PostScript code which adds the ImScaleDecode
filter to the DataSource, but there seems little point in doing this in
C when it could be done more easily in PostScript.

So instead I've chosen to test if the current device is a 'high level'
device and, if it is, not to apply the filter.

With pdfwrite this leads to us producing smaller PDF files, as expected,
and the result in Acrobat matching the one from Distiller.

When testing with the cluster a small number of Genoa test files exhibit
differences. These all use the well known 'turkey' image at very low resolution,
with the old code these were interpolated at 720 dpi and that was the image
data written out. Now we write the original data and interpolate to the
resolution of the actual device at rendering time. In the case of 300 dpi
output this results in small but visible differences.