Bug 694561: Fix interpolated landscape images (again).
Don't try to accumulate runs of identical pixels in landscape images.
We weren't aligning the transposed data to bitmap_align, and this was
causing problems. It's unfeasible to do so, and the underlying driver
almost certainly will break down to 1 call per pixel, so we might as
well do that immediately and simplify the code massively.

Also fix some warnings/unused variables.