Bug 690790: Workaround libpng 64bit build bug.
libpng keeps a jmpbuf in it's structure. jmp_bufs have to be 16 byte
aligned on x64 systems, and the png allocator makes no effort to
align the buffer.

The workaround here changes the PNG allocator to allocate it's main
structure in a larger block than it needs, and to actually use a
block in the middle of this to ensure alignment. A pointer is
kept to the real address of the block for freeing.