PDF interpreter - remove the 2D array of objects for large (>64k) objects
Bug #694470

In the past Ghostscript had a limit of 64k entries for arrays and strings
but this proved a problem for some PDF files, so a set of new objects were
defined 'larray' and 'lstring', which were defined as an array of arrays
and an array of strings. New accessor and utility functions were also
created.

With bug 694470 there is an object with an index of 2^32 - 1, which takes
a very long time, as the code allocates more and more sub-arrays.

Now that we have a limit of 2^24 on arrays and strings (which can be increased),
we can remove the 2d array hackery. This causes the bug file to abort almost
immediately while rebuilding the xref. It *should* also provide a modest
performance benefit, and makes the code smaller and easier to deal with.

No differences expected.