Bug 694590: Tweak scan converters handling of horizontal lines.
The scan converter only puts horizontal edges into the edge list
if expanding the height of the line by fill adjust causes it to
cover a pixel centre.

Unfortunately, for the fill_adjust = 0.5 case we actually use
adjust_above = 128 and adjust_below = 127, so there is one possible
subpixel offset for which the line will not appear. We therefore
tweak the logic to accept all horizontal lines through in this case.

This produces many diffs, all benign. Sorry Marcos.