pdfwrite - add support for form preservation from PostScript
Bug #687561 et al

We add the new operators .beginform and .endform which call device spec_op
methods in order to allow devices to capture form streams. Devices can
also return a matrix to replace the CTM for the course of the form.

Modified pdfwrite to capture and store forms, and set the CTM to the
pdfwrite general matrix (resolution/72).

Added new PDFexecform for use by the PDF interpreter, trying to capture
forms from PDF files is fraught with difficulty because we already capture
transparent forms as forms.

Because some kinds of forms cannot be successfully preserved (only happens
with badly behaved PostScript) we need a way to prevent forms from being
preserved. This option does that.

Examples of badly behaved PostScript can be found in a few Quality Logic
CET tests (use of setgstate in a form to set the CTM to a state prior to
the form execution, and use of setflat with a parameter derived from the
CTM. So we add UNROLLFORMS to the gs_cet file.